Relax mimemagic version requirement
The mimemagic gem included a file that was taken from a GPL-licensed
project. That meant that mimemagic had to use the GPL license too, but
it was using the MIT license at the time. The author responded by
removing versions 0.3.5 and earlier from rubygems.

We currently depend on paperclip version 4.3.7, which has pinned
mimemagic version 0.3.0. Since we currently cannot upgrade paperclip,
and mimemagic 0.3.0 is no longer available, we have to fork the gem and
relax the requirement. Later versions of paperclip are less strict about
the patch version of mimemagic, so it's possible this dependency
requirement was overly strict.